+++ src/sentry/testutils/cases.py
@@ -51,7 +51,7 @@ from django.core.cache import cache
 from django.db import DEFAULT_DB_ALIAS, connection, connections
 from django.db.migrations.executor import MigrationExecutor
 from django.http import HttpRequest
-from django.test import TestCase, TransactionTestCase, override_settings
+from django.test import LiveServerTestCase, TestCase, TransactionTestCase, override_settings
 from django.test.utils import CaptureQueriesContext
 from django.urls import reverse
 from django.utils import timezone
@@ -370,6 +370,15 @@ class TestCase(BaseTestCase, TestCase):
 
 class TransactionTestCase(BaseTestCase, TransactionTestCase):
     pass
+
+
+class LiveServerTestCase(BaseTestCase, LiveServerTestCase):
+    def setUp(self):
+        super().setUp()
+
+        opt = self.options({"system.url-prefix": self.live_server_url})
+        opt.__enter__()
+        self.addCleanup(opt.__exit__, None, None, None)
 
 
 class APITestCase(BaseTestCase, BaseAPITestCase):
